[INLONG-5259][Agent] File data sources support custom end-of-line symbols
