+++ inlong-agent/agent-plugins/pom.xml
@@ -139,9 +139,5 @@
             <artifactId>powermock-api-mockito2</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>io.fabric8</groupId>
-            <artifactId>kubernetes-client</artifactId>
-        </dependency>
     </dependencies>
 </project>
