+++ inlong-agent/agent-plugins/src/main/java/org/apache/inlong/agent/plugin/sources/TextFileSource.java
@@ -18,21 +18,14 @@
 package org.apache.inlong.agent.plugin.sources;
 
 import org.apache.inlong.agent.conf.JobProfile;
-import org.apache.inlong.agent.constant.FileConstants;
-import org.apache.inlong.agent.constant.JobConstants;
 import org.apache.inlong.agent.plugin.Reader;
 import org.apache.inlong.agent.plugin.Source;
-import org.apache.inlong.agent.plugin.SourceMeta;
 import org.apache.inlong.agent.plugin.sources.reader.TextFileReader;
-import org.apache.inlong.agent.plugin.utils.MetaDataUtils;
 import org.apache.inlong.agent.plugin.utils.PluginUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
-import java.io.FileReader;
-import java.io.IOException;
-import java.io.LineNumberReader;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -56,7 +49,6 @@ public class TextFileSource implements Source {
     // path + suffix
     private static final Logger LOGGER = LoggerFactory.getLogger(TextFileSource.class);
     private static final String TEXT_FILE_SOURCE_TAG_NAME = "AgentTextFileSourceMetric";
-    private List<SourceMeta> sourceMetas;
 
     public TextFileSource() {
     }
@@ -70,10 +62,10 @@ public class TextFileSource implements Source {
         List<Reader> result = new ArrayList<>();
         String filterPattern = jobConf.get(JOB_LINE_FILTER_PATTERN, DEFAULT_JOB_LINE_FILTER);
         for (File file : allFiles) {
-            int seekPosition = getSeekPosition(jobConf, file);
+            int seekPosition = jobConf.getInt(file.getAbsolutePath() + POSITION_SUFFIX, 0);
             LOGGER.info("read from history position {} with job profile {}, file absolute path: {}", seekPosition,
                     jobConf.getInstanceId(), file.getAbsolutePath());
-            TextFileReader textFileReader = new TextFileReader(file, seekPosition, sourceMetas);
+            TextFileReader textFileReader = new TextFileReader(file, seekPosition);
             long waitTimeout = jobConf.getLong(JOB_READ_WAIT_TIMEOUT, DEFAULT_JOB_READ_WAIT_TIMEOUT);
             textFileReader.setWaitMillisecond(waitTimeout);
             addValidator(filterPattern, textFileReader);
@@ -82,28 +74,6 @@ public class TextFileSource implements Source {
         // increment the count of successful sources
         GLOBAL_METRICS.incSourceSuccessCount(metricTagName);
         return result;
-    }
-
-    @Override
-    public void initSourceMeta(JobProfile jobProfile) {
-        this.sourceMetas = MetaDataUtils.getSourceMeta(jobProfile);
-    }
-
-    private int getSeekPosition(JobProfile jobConf, File file) {
-        int seekPosition;
-        if (jobConf.hasKey(JobConstants.JOB_FILE_COLLECT_TYPE) && FileConstants.INCREMENT.name()
-                .equalsIgnoreCase(jobConf.get(JobConstants.JOB_FILE_COLLECT_TYPE))) {
-            try (LineNumberReader lineNumberReader = new LineNumberReader(new FileReader(file.getPath()))) {
-                lineNumberReader.skip(Long.MAX_VALUE);
-                seekPosition = lineNumberReader.getLineNumber() + 1;
-                return seekPosition;
-            } catch (IOException ex) {
-                LOGGER.error("get position error, file absolute path: {}", file.getAbsolutePath());
-                throw new RuntimeException(ex);
-            }
-        }
-        seekPosition = jobConf.getInt(file.getAbsolutePath() + POSITION_SUFFIX, 0);
-        return seekPosition;
     }
 
     private void addValidator(String filterPattern, TextFileReader textFileReader) {
