+++ inlong-agent/agent-plugins/src/test/resources/test/1.txt
@@ -1,3 +1,3 @@
-hello bb aa
-world bb
-agent bb
+hello line-end-symbol aa
+world line-end-symbol
+agent line-end-symbol
