+++ inlong-agent/agent-plugins/src/test/resources/test/2.txt
@@ -1,2 +1,2 @@
 hello
-world bb
+world line-end-symbol
