+++ inlong-agent/agent-plugins/src/test/resources/test/a.txt
@@ -1,3 +1,3 @@
-test bb a
-test b bb
+test line-end-symbol a
+test b line-end-symbol
 agent
