+++ pom.xml
@@ -250,8 +250,6 @@
         <HikariCP.version>4.0.3</HikariCP.version>
 
         <kafka.clients.version>3.0.0</kafka.clients.version>
-        
-        <kubernetes.client.version>6.0.0</kubernetes.client.version>
     </properties>
 
     <dependencyManagement>
@@ -1554,11 +1552,6 @@
                 <groupId>com.qcloud</groupId>
                 <artifactId>dlc-data-catalog-metastore-client</artifactId>
                 <version>${dlc.client.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>io.fabric8</groupId>
-                <artifactId>kubernetes-client</artifactId>
-                <version>${kubernetes.client.version}</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
